refactor typed signature
