+++ app/components/UI/SignatureRequest/SignatureSharedState/index.ts
@@ -1,0 +1,2 @@
+export { default as signatureAnalytics } from './signatureAnalytics';
+export { default as showWalletConnectNotification } from './showWCNotification';
